# chore: speed up linting