--- a/browser-test/tsconfig.build.json
+++ b/browser-test/tsconfig.build.json
@@ -1,7 +1,7 @@
 {
   "extends": [
     "./tsconfig.json",
-    "../../tsconfig-build-options.json"
+    "../tsconfig-build-options.json"
   ],
   "exclude": [
     "test/"