--- a/browser-test/tsconfig.json
+++ b/browser-test/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../tsconfig.eslint-base.json",
+  "include": [
+    "*.js",
+    "tests/*.js"
+  ]
+}