--- a/package.json
+++ b/package.json
@@ -35,8 +35,13 @@
     "docs:markdown-for-agoric-documentation-repo": "typedoc --plugin typedoc-plugin-markdown --tsconfig tsconfig.build.json",
     "update": "echo 'DEPRECATED: use \"yarn upgrade-interactive\" or \"yarn dedupe\"'",
     "format": "prettier --write .github packages",
-    "lint": "prettier --check .github packages && yarn workspaces foreach --all run lint",
-    "lint-fix": "yarn workspaces foreach --all run lint-fix",
+    "lint": "yarn lint:prettier && yarn lint:eslint",
+    "lint:eslint": "eslint .",
+    "lint:workspaces": "prettier --check .github packages && yarn workspaces foreach --all run lint",
+    "lint:workspaces:eslint": "yarn workspaces foreach --all run lint:eslint",
+    "lint:fix": "yarn lint:eslint -- --fix",
+    "lint:workspaces:fix": "yarn workspaces foreach --all run lint-fix",
+    "lint:prettier": "prettier --check .github packages",
     "test": "yarn workspaces foreach --all --exclude @endo/skel run test",
     "test:xs": "yarn workspaces foreach --all run test:xs",
     "test262": "yarn workspaces foreach --all run test262",
@@ -56,5 +61,8 @@
   "resolutions": {
     "@lerna/version@npm:5.6.2": "patch:@lerna/version@npm%3A5.6.2#~/.yarn/patches/@lerna-version-npm-5.6.2-ce2d9cb2f5.patch",
     "@lerna/conventional-commits@npm:5.6.2": "patch:@lerna/conventional-commits@npm%3A5.6.2#~/.yarn/patches/@lerna-conventional-commits-npm-5.6.2-a373ba4bc0.patch"
+  },
+  "eslintConfig": {
+    "root": true
   }
 }