--- a/packages/bundle-source/.eslintignore
+++ b/packages/bundle-source/.eslintignore
@@ -1,3 +1,0 @@
-/dist/
-/scripts/
-/demo/