--- a/packages/bundle-source/package.json
+++ b/packages/bundle-source/package.json
@@ -76,6 +76,11 @@
   "eslintConfig": {
     "extends": [
       "plugin:@endo/internal"
+    ],
+    "ignorePatterns": [
+      "dist/",
+      "scripts/",
+      "demo/"
     ]
   },
   "typeCoverage": {