--- a/packages/bundle-source/test/explicit.test.js
+++ b/packages/bundle-source/test/explicit.test.js
@@ -5,6 +5,7 @@ import test from '@endo/ses-ava/prepare-endo.js';
 import bundleSource from '../src/index.js';
 
 test('explicit authority', async t => {
+  // @ts-expect-error BundleOptions needs updating
   const { moduleFormat } = await bundleSource(
     url.fileURLToPath(new URL(`../demo/dir1`, import.meta.url)),
     'getExport',