--- a/packages/bundle-source/test/external-fs.test.js
+++ b/packages/bundle-source/test/external-fs.test.js
@@ -37,6 +37,7 @@ const testFsImportHookEndoZipBase64 = (name, file) => {
 
     const testFile = url.fileURLToPath(new URL(file, import.meta.url));
 
+    // @ts-expect-error BundleOptions needs updating
     await bundleSource(testFile, {
       format: 'endoZipBase64',
       importHook: async specifier => {