--- a/packages/eslint-plugin/lib/index.js
+++ b/packages/eslint-plugin/lib/index.js
@@ -2,18 +2,19 @@
  * @module Agoric-specific plugin
  * @author Agoric
  */
+
 'use strict';
 
 //------------------------------------------------------------------------------
 // Requirements
 //------------------------------------------------------------------------------
 
-var requireIndex = require('requireindex');
+const requireIndex = require('requireindex');
 
 //------------------------------------------------------------------------------
 // Plugin Definition
 //------------------------------------------------------------------------------
 
 // import all rules in lib/rules
-module.exports.rules = requireIndex(__dirname + '/rules');
-module.exports.configs = requireIndex(__dirname + '/configs');
+module.exports.rules = requireIndex(`${__dirname}/rules`);
+module.exports.configs = requireIndex(`${__dirname}/configs`);