--- a/packages/eslint-plugin/lib/rules/no-polymorphic-call.js
+++ b/packages/eslint-plugin/lib/rules/no-polymorphic-call.js
@@ -1,3 +1,5 @@
+/* eslint-disable no-use-before-define */
+
 'use strict';
 
 module.exports = {