--- a/packages/init/package.json
+++ b/packages/init/package.json
@@ -68,11 +68,11 @@
   "eslintConfig": {
     "extends": [
       "plugin:@endo/internal"
+    ],
+    "ignorePatterns": [
+      "bundle-*.js"
     ]
   },
-  "eslintIgnore": [
-    "bundle-*.js"
-  ],
   "publishConfig": {
     "access": "public"
   },