--- a/packages/module-source/package.json
+++ b/packages/module-source/package.json
@@ -78,6 +78,9 @@
   "eslintConfig": {
     "extends": [
       "plugin:@endo/internal"
+    ],
+    "ignorePatterns": [
+      "test/fixtures/"
     ]
   },
   "ava": {