--- a/packages/skel/tsconfig.build.json
+++ b/packages/skel/tsconfig.build.json
@@ -3,9 +3,6 @@
     "./tsconfig.json",
     "../../tsconfig-build-options.json"
   ],
-  "compilerOptions": {
-    "allowJs": true
-  },
   "exclude": [
     "test/"
   ]