--- a/packages/test262-runner/.eslintignore
+++ b/packages/test262-runner/.eslintignore
@@ -1,1 +1,0 @@
-prelude