--- a/packages/test262-runner/package.json
+++ b/packages/test262-runner/package.json
@@ -48,6 +48,9 @@
   "eslintConfig": {
     "extends": [
       "plugin:@endo/internal"
+    ],
+    "ignorePatterns": [
+      "prelude/"
     ]
   },
   "files": [