--- a/packages/test262-runner/tsconfig.json
+++ b/packages/test262-runner/tsconfig.json
@@ -9,5 +9,6 @@
     "scripts",
     "src",
     "test"
-  ]
+  ],
+  "exclude": ["prelude"]
 }