Previously, linting took this long (M4 mac):

```
yarn lint:workspaces:eslint  88.01s user 8.69s system 165% cpu 58.326 total
```

With this change:

```
yarn lint:eslint  18.74s user 2.13s system 166% cpu 12.545 total
```

## Description

This PR tweaks ESLint and TypeScript config to allow the entire monorepo to be linted from the monorepo root:

```shell
eslint .
```

Summary of changes:

- Add a root `.eslintignore` file, since it's kind of a lot to throw in `package.json`
- Configure the workspace root as the ESLint root, which prevents ESLint from attempting to traverse past it looking for configuration
- Move some small `.eslintignore` files into `package.json`-based ESLint config
- Add TS configuration files to workspaces missing them, since tseslint expects them to be there
- Removed redundant `allowJs: true` and `checkJs: true` in various TS config files (including `skel/tsconfig.build.json`; did not check _all_ of them)
- Removed unused `eslintIgnore` field from `lockdown/package.json`
- Migrated `eslintIgnore` field to `eslintConfig.ignorePatterns` in `init/package.json`
- Enable lint of the ESLint plugin 😄
- Shuffle scripts around; `yarn lint:eslint` now uses the new strategy; `yarn lint:workspaces:eslint` uses the old strategy (if we want to keep it)

---

The second commit applies lint fixes to previously-unlinted files.