@boneskull 2025-07-03
@kriskowal 

> Does the story for linting a package in isolation change?

Each package can be linted in isolation.  This is what the `lint:workspaces:eslint` script does.

Similarly, any file or subdirectory can be linted via:

```sh
npx eslint <glob>
```

It'd make sense _to me_ to keep all common dev deps in the workspace root `package.json`, but IDK how well Yarn supports that strategy (so I didn't try to change it).  If it works the way it should, practically speaking, there won't be any change to how tools get run (because they all wind up in `/node_modules/.bin` anyhow).  It _does_ mean that a workspace cannot be