@boneskull 2025-07-15
@kriskowal FWIW, if we _do_ end up moving `devDependencies` into the workspace root, then any tool installed this way can always be executed from any workspace or subdirectory via `npx` or even `npm exec`.  There may be a `yarn` equivalent, but those two commands shouldn't conflict with Yarn in any way.